Split text diffing from tree diffing
